Moved pagination code to mixin
